Unlink client with DataStore

+++ components/brave_federated/linear_algebra_util/linear_algebra_util.cc
@@ -1,3 +1,8 @@
+/* Copyright (c) 2022 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 #include "brave/components/brave_federated/linear_algebra_util/linear_algebra_util.h"
 
 #include <vector>
@@ -7,7 +12,7 @@ namespace brave_federated {
 std::vector<float> LinearAlgebraUtil::SubtractVector(std::vector<float> v1,
                                                      std::vector<float> v2) {
   std::vector<float> result(v1.size());
-  for (int i = 0; i < (int)v1.size(); i++) {
+  for (size_t i = 0; i < v1.size(); i++) {
     result[i] = v1[i] - v2[i];
   }
 
@@ -18,9 +23,9 @@ std::vector<float> LinearAlgebraUtil::MultiplyMatrixVector(
     std::vector<std::vector<float>> mat,
     std::vector<float> v) {
   std::vector<float> result(mat.size(), 0.0);
-  for (int i = 0; i < (int)mat.size(); i++) {
+  for (size_t i = 0; i < mat.size(); i++) {
     result[i] = 0;
-    for (int j = 0; j < (int)mat[0].size(); j++) {
+    for (size_t j = 0; j < mat[0].size(); j++) {
       result[i] += mat[i][j] * v[j];
     }
   }
@@ -30,7 +35,7 @@ std::vector<float> LinearAlgebraUtil::MultiplyMatrixVector(
 
 std::vector<float> LinearAlgebraUtil::AddVectorScalar(std::vector<float> v,
                                                       float a) {
-  for (int i = 0; i < (int)v.size(); i++) {
+  for (size_t i = 0; i < v.size(); i++) {
     v[i] += a;
   }
 
@@ -38,8 +43,8 @@ std::vector<float> LinearAlgebraUtil::AddVectorScalar(std::vector<float> v,
 }
 
 std::vector<float> LinearAlgebraUtil::AddVectors(std::vector<float> v1,
-                                                      std::vector<float> v2) {
-  for (int i = 0; i < (int)v1.size(); i++) {
+                                                 std::vector<float> v2) {
+  for (size_t i = 0; i < v1.size(); i++) {
     v1[i] += v2[i];
   }
 
@@ -48,7 +53,7 @@ std::vector<float> LinearAlgebraUtil::AddVectors(std::vector<float> v1,
 
 std::vector<float> LinearAlgebraUtil::MultiplyVectorScalar(std::vector<float> v,
                                                            float a) {
-  for (int i = 0; i < (int)v.size(); i++) {
+  for (size_t i = 0; i < v.size(); i++) {
     v[i] *= a;
   }
 
@@ -58,28 +63,30 @@ std::vector<float> LinearAlgebraUtil::MultiplyVectorScalar(std::vector<float> v,
 std::vector<std::vector<float>> LinearAlgebraUtil::MultiplyMatrices(
     std::vector<std::vector<float>> mat1,
     std::vector<std::vector<float>> mat2) {
-      std::vector<std::vector<float>> result(mat1.size(), std::vector<float>(mat2[0].size(), 0.0));
+  std::vector<std::vector<float>> result(
+      mat1.size(), std::vector<float>(mat2[0].size(), 0.0));
 
-  for (int i = 0; i < (int) mat1.size(); i++) {
-    for (int j = 0; j < (int) mat2[0].size(); j++) {
-      for (int k = 0; k < (int) mat1[0].size(); k++) {
+  for (size_t i = 0; i < mat1.size(); i++) {
+    for (size_t j = 0; j < mat2[0].size(); j++) {
+      for (size_t k = 0; k < mat1[0].size(); k++) {
         result[i][j] += mat1[i][k] * mat2[k][j];
       }
     }
   }
-  
+
   return result;
 }
 
 std::vector<std::vector<float>> LinearAlgebraUtil::TransposeVector(
     std::vector<std::vector<float>> v) {
   std::vector<std::vector<float>> vT(v[0].size(), std::vector<float>(v.size()));
-  for (int i = 0; i < (int)v.size(); i++) {
-    for (int j = 0; j < (int)v[0].size(); j++) {
+  for (size_t i = 0; i < v.size(); i++) {
+    for (size_t j = 0; j < v[0].size(); j++) {
       vT[j][i] = v[i][j];
     }
   }
 
   return vT;
 }
+
 }  // namespace brave_federated
